tests/wpt/tests/encoding-detection/ar-ISO-8859-6.tentative.html

<!doctype html>
<title>ar ISO-8859-6</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<p>Á–« ÁË «Œ »«—  —ÂÍ“ «‰√Õ—·.</p>
<script>
setup({explicit_done:true});
onload = function() {
    test(function() {
        assert_equals(document.characterSet, "ISO-8859-6", 'Expected ISO-8859-6');
    }, "Check detection result");
      done();
};
</script>
