third_party/blink/web_tests/platform/generic/external/wpt/encoding-detection/ar-ISO-8859-6.tentative-expected.txt

This is a testharness.js-based test.
FAIL Check detection result assert_equals: Expected ISO-8859-6 expected "ISO-8859-6" but got "windows-1256"
Harness: the test ran to completion.
